+++ src/components/buttons/default/button.stories.js
@@ -2,9 +2,14 @@ import { storiesOf } from '@storybook/vue'
 
 storiesOf('Buttons/default', module)
   .add('primary', () => ({
+    methods: {
+      handleClick (event) {
+        console.log(event)
+      }
+    },
     template: `
       <div>
-        <w-button color="primary">
+        <w-button color="primary" @click="handleClick">
           Click me
         </w-button>
       </div>