+++ src/components/buttons/flat/button-flat.vue
@@ -5,6 +5,7 @@
     :disabled="disabled"
     :size="size"
     :style="style"
+    @click="$emit('click', $event)"
   >
     <w-typography variant="button">
       <slot />