+++ src/components/icon/icon.stories.js
@@ -1,0 +1,8 @@
+import { storiesOf } from '@storybook/vue'
+
+storiesOf('Icon', module)
+  .add('default', () => ({
+    template: `
+     <w-icon name="file" />
+    `
+  }))